# demo/src/main/resources/static/openapi.yaml
openapi: 3.0.1
servers:
  # Added by API Auto Mocking Plugin
  - description: localhost
    url: http://localhost:8080
info:
  version: "1.0.0"
  title: swiss-api
...
paths:
...
  /validation/{email}/{code}:
    post:
      tags:
        - Validation
      operationId: validateEmail
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
        - name: code
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: code successfully validated
          content:
            application/json:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
  /validation/forgetpassword:
    post:
      tags:
        - Validation
      operationId: forgetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgetPassword'
      responses:
        '200':
          description: email was sent
          content:
            application/json:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
  /validation/forgetpasswordrequest/{email}:
    post:
      tags:
        - Validation
      operationId: forgetPasswordRequest
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: email was sent
          content:
            application/json:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'

  #Support
...
components:
  schemas:
...
    ForgetPassword:
      type: object
      description: you can use it for forget pswd
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        passwordConfirmation:
          type: string
          format: password
...
    ClientErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ClientError'
      type: object
      example:
        error:
          message: 'Authorization information (Header: %s) is missing or invalid.'
...
    ServerErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ServerError'
      type: object
      example:
        error:
          message: 'Internal Server Error.'
          trace_id: 66
...
    TooManyRequestsResponse:
      properties:
        error:
          $ref: '#/components/schemas/ClientError'
      type: object
      example:
        error:
          message: 'Too many requests'
...
    ClientError:
      properties:
        message:
          type: string
      type: object
...
    ServerError:
      properties:
        message:
          type: string
        trace_id:
          description: 'You should send us this ID to investigate the issue.'
          type: string
      type: object
...
  responses:
    BadRequest:
      description: 'The request is malformed.'
      headers:
        X-RateLimit-Limit:
          description: 'Request limit per minute.'
          schema:
            type: integer
            format: int32
        X-RateLimit-Remaining:
          description: 'The number of requests left for the time window.'
          schema:
            type: integer
            format: int32
        X-RateLimit-Reset:
          description: 'The timestamp at which the current rate limit window resets.'
          schema:
            type: integer
            format: int32
        Retry-After:
          description: 'How many seconds you have to wait before making new request.'
          schema:
            type: integer
            format: int32
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClientErrorResponse'
    Unauthorized:
      description: 'Authorization information is missing or invalid.'
      headers:
        X-RateLimit-Limit:
          description: 'Request limit per minute.'
          schema:
            type: integer
            format: int32
        X-RateLimit-Remaining:
          description: 'The number of requests left for the time window.'
          schema:
            type: integer
            format: int32
        X-RateLimit-Reset:
          description: 'The timestamp at which the current rate limit window resets.'
          schema:
            type: integer
            format: int32
        Retry-After:
          description: 'How many seconds you have to wait before making new request.'
          schema:
            type: integer
            format: int32
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClientErrorResponse'
...
    NotFound:
      description: 'Non-existent resource is requested.'
      headers:
        X-RateLimit-Limit:
          description: 'Request limit per minute.'
          schema:
            type: integer
            format: int32
        X-RateLimit-Remaining:
          description: 'The number of requests left for the time window.'
          schema:
            type: integer
            format: int32
        X-RateLimit-Reset:
          description: 'The timestamp at which the current rate limit window resets.'
          schema:
            type: integer
            format: int32
        Retry-After:
          description: 'How many seconds you have to wait before making new request.'
          schema:
            type: integer
            format: int32
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClientErrorResponse'
...
    ServerError:
      description: 'Internal server error.'
      headers:
        X-RateLimit-Limit:
          description: 'Request limit per minute.'
          schema:
            type: integer
            format: int32
        X-RateLimit-Remaining:
          description: 'The number of requests left for the time window.'
          schema:
            type: integer
            format: int32
        X-RateLimit-Reset:
          description: 'The timestamp at which the current rate limit window resets.'
          schema:
            type: integer
            format: int32
        Retry-After:
          description: 'How many seconds you have to wait before making new request.'
          schema:
            type: integer
            format: int32
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ServerErrorResponse'
    TooManyRequests:
      description: 'Too many requests'
      headers:
        X-RateLimit-Limit:
          description: 'Request limit per minute.'
          schema:
            type: integer
            format: int32
        X-RateLimit-Remaining:
          description: 'The number of requests left for the time window.'
          schema:
            type: integer
            format: int32
        X-RateLimit-Reset:
          description: 'The timestamp at which the current rate limit window resets.'
          schema:
            type: integer
            format: int32
        Retry-After:
          description: 'How many seconds you have to wait before making new request.'
          schema:
            type: integer
            format: int32
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsResponse'
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes